small performance tweaks

+++ src/render/fragment.glsl
@@ -1,5 +1,5 @@
 varying lowp vec4 vColor;
-varying highp vec3 vLighting;
+varying lowp vec3 vLighting;
 
 void main(void) {
 	gl_FragColor = vColor * vec4(vLighting, 1);
